refactor: update seeding functions with new logger
