+++ src/controllers/groups/index.test.ts
@@ -1,7 +1,7 @@
 import { OAuthProvider, ReputationLevel } from "@interrep/reputation"
 import { appendLeaf } from "src/core/groups/mts"
 import createNextMocks from "src/mocks/createNextMocks"
-import seedZeroHashes from "src/utils/backend/seeding/seedZeroHashes"
+import { seedZeroHashes } from "src/utils/backend/seeding"
 import { clearDatabase, connectDatabase, dropDatabaseAndDisconnect } from "src/utils/backend/testDatabase"
 import getGroup from "./getGroup"
 import getGroups from "./getGroups"
