+++ src/core/groups/index.test.ts
@@ -3,7 +3,7 @@ import { poseidon } from "circomlibjs"
 import { appendLeaf } from "src/core/groups/mts"
 import { PoapEvent } from "src/core/poap"
 import { Web3Provider } from "src/types/groups"
-import seedZeroHashes from "src/utils/backend/seeding/seedZeroHashes"
+import { seedZeroHashes } from "src/utils/backend/seeding"
 import { clearDatabase, connectDatabase, dropDatabaseAndDisconnect } from "src/utils/backend/testDatabase"
 import { checkGroup, getGroup, getGroups, getProviders } from "."
 
@@ -44,7 +44,7 @@ describe("# core/groups", () => {
         })
 
         it("Should return the group with size = 10", async () => {
-            await seedZeroHashes(false)
+            await seedZeroHashes()
 
             for (let i = 0; i < 10; i++) {
                 const idCommitment = poseidon([BigInt(i)]).toString()
