+++ src/utils/backend/seeding/index.ts
@@ -1,0 +1,4 @@
+import seedZeroHashes from "./seedZeroHashes"
+import seedTwitterUsers from "./seedTwitterUsers"
+
+export { seedZeroHashes, seedTwitterUsers }
