+++ src/utils/backend/seeding/seedTwitterUsers.ts
@@ -1,52 +1,27 @@
-import { ReputationLevel, OAuthProvider } from "@interrep/reputation"
-import colors from "colors"
 import { OAuthAccount } from "@interrep/db"
-import { getTwitterFriendsByUserId, getTwitterUserByUsername } from "src/services/twitter"
-import { TwitterUser } from "src/types/twitter"
+import { OAuthProvider, ReputationLevel } from "@interrep/reputation"
+import { getTwitterUserByUsername } from "src/services/twitter"
+import logger from "src/utils/backend/logger"
 
-function createTwitterSeedUser(user: TwitterUser): any {
-    return new OAuthAccount({
-        provider: OAuthProvider.TWITTER,
-        providerAccountId: user.id,
-        reputation: ReputationLevel.GOLD
-    })
-}
-
-export default async function seedTwitterUsers(twitterUsernames: string[], logger = false): Promise<void> {
-    const log = logger ? console.log : (message: string) => message
-
-    log(colors.white.bold("Seeding Twitter accounts...\n"))
+export default async function seedTwitterUsers(twitterUsernames: string[]): Promise<void> {
+    logger.verbose("Seeding Twitter accounts...")
 
     for (const username of twitterUsernames) {
         const twitterUser = await getTwitterUserByUsername({
             username
         })
-        const friends = await getTwitterFriendsByUserId({
-            userId: twitterUser.id,
-            maxResults: 900
-        })
+        const account = await OAuthAccount.findByProviderAccountId(OAuthProvider.TWITTER, twitterUser.id)
 
-        log(colors.white(`${username} has ${friends.length} friends`))
-
-        if (friends.length === 0) {
-            break
-        }
-
-        const formattedFriends = friends.map((friend) => createTwitterSeedUser(friend))
-
-        try {
-            // With ordered false, it inserts all documents it can and report
-            // errors at the end (incl. errors from duplicates).
-            const docs = await OAuthAccount.insertMany(formattedFriends, {
-                ordered: false
+        if (!account) {
+            await OAuthAccount.create({
+                provider: OAuthProvider.TWITTER,
+                providerAccountId: twitterUser.id,
+                reputation: ReputationLevel.GOLD
             })
 
-            log(colors.green.bold(`${docs.length} ${username}'s friends inserted ✓`))
-        } catch (error: any) {
-            log(colors.white(`${error.result?.nInserted} documents have been inserted`))
-            log(colors.white(`Number of write errors: ${error.writeErrors?.length}`))
+            logger.verbose(`Twitter user '${username}' has been inserted`)
         }
-
-        log("")
     }
+
+    logger.info(`All the Twitter users have been inserted correctly`)
 }
