+++ src/utils/backend/seeding/seedZeroHashes.ts
@@ -1,12 +1,10 @@
-import config from "src/config"
-import colors from "colors"
 import { MerkleTreeZero } from "@interrep/db"
+import config from "src/config"
+import logger from "src/utils/backend/logger"
 import { poseidon } from "src/utils/common/crypto"
 
-export default async function seedZeroHashes(logger = false): Promise<void> {
-    const log = logger ? console.log : (message: string) => message
-
-    log(colors.white.bold("Seeding zero hashes...\n"))
+export default async function seedZeroHashes(): Promise<void> {
+    logger.verbose("Seeding zero hashes...")
 
     let level = 0
     let zeroHash = "0"
@@ -14,26 +12,20 @@ export default async function seedZeroHashes(logger = false): Promise<void> {
     const zeroHashes = await MerkleTreeZero.find()
 
     if (zeroHashes && zeroHashes.length > 0) {
-        log(colors.white(`There are already ${zeroHashes.length} zero hashes!\n`))
-
         level = zeroHashes.length
         zeroHash = zeroHashes[level - 1].hash
     }
 
-    if (level < config.MERKLE_TREE_DEPTH) {
-        for (level; level < config.MERKLE_TREE_DEPTH; level++) {
-            zeroHash = level === 0 ? zeroHash : poseidon(zeroHash, zeroHash)
+    for (level; level < config.MERKLE_TREE_DEPTH; level++) {
+        zeroHash = level === 0 ? zeroHash : poseidon(zeroHash, zeroHash)
 
-            const zeroHashDocument = await MerkleTreeZero.create({
-                level,
-                hash: zeroHash
-            })
+        await MerkleTreeZero.create({
+            level,
+            hash: zeroHash
+        })
 
-            await zeroHashDocument.save()
-
-            log(colors.white(`Document with id: ${zeroHashDocument.id} inserted`))
-        }
-
-        log(colors.green.bold("\nDocuments inserted correctly ✓\n"))
+        logger.verbose(`Zero hash '${zeroHash}' has been inserted`)
     }
+
+    logger.info("All the zero hashes have been inserted correctly")
 }
